Login form/authentication for the new devtools
You need to use wrangler and the account_management krl files in order for the oauth side of this to work. In order to see the oauth picos etc in effect in new devtools, you must also make three changes to the index.js file to visually see the picos.