basic engine discovery.
Seven seconds after the engine starts a discovery service is started.
discover service allows you to add resources to be advertised to other pico engines on the network.  discover service also allows you to register observer DiD's. when discover finds another engine, it will send an event to all the observer dids with what that engine is advertising. This enables any observer to listen for different types of resources and the ability to subscribe to them.
The dependency has not been added. you will need to install node-discover manually