new rabbit hole.
The structure for code feels correct at this point. That is discover.js lives in pico-engine and not pico-engine core. The problem at this point is I need to be able to pass in a handle to the pico-engine during the construction of discover, while providing a way to export module functions at pico-engine-core creation.  I think the way we allow custom defined JavaScript modules may need to evolve for this scenario.   discover.js needs to be refactor into a class with a constructor.